requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
Flask-Login==0.6.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
